Se implementan y corrigen entities, services y dtos

+++ src/orders/services/order-item.service.ts
@@ -4,7 +4,10 @@ import { Repository } from 'typeorm';
 import { OrderItem } from '@ordersModule/entities/order-item.entity';
 import { Product } from '@productsModule/entities/product.entity';
 import { Order } from '@ordersModule/entities/order.entity';
-import { createOrderItemDto, updateOrderItemDto } from '@ordersModule/dtos/order-item.dto';
+import {
+  createOrderItemDto,
+  updateOrderItemDto,
+} from '@ordersModule/dtos/order-item.dto';
 
 @Injectable()
 export class OrderItemService {
@@ -16,30 +19,52 @@ export class OrderItemService {
     @InjectRepository(Product)
     private productsRepository: Repository<Product>,
   ) {}
-  async create (payload: createOrderItemDto) {
-    const order = await this.ordersRepository.findOne({where: {id: payload.idOrder}});
-    const product = await this.productsRepository.findOne({where: {id: payload.idProducto}});
+  async create(payload: createOrderItemDto) {
+    const order = await this.ordersRepository.findOne({
+      where: { id: payload.idOrder },
+    });
+    const product = await this.productsRepository.findOne({
+      where: { id: payload.idProducto },
+    });
 
     const orderItem = this.orderItemRepository.create({
       quantity: payload.quantity,
       product,
-      order
+      order,
     });
     return this.orderItemRepository.save(orderItem);
   }
 
-  async delete(orderId: number, productId: number) {
-    const item = await this.orderItemRepository.findOne({
+  async update(idOrdenItem: number, payload: updateOrderItemDto) {
+    const orderItem = await this.orderItemRepository.findOne({
       where: {
-        order: {
-          id: orderId
-        },
-        product: {
-          id: productId
-        }
-      }
+        id: idOrdenItem,
+      },
     });
-    return await this.orderItemRepository.remove(item);
+    if (!orderItem) return null;
+    const product = await this.productsRepository.findOne({
+      where: { id: payload.idProducto },
+    });
+    if (!product) return null;
+    const order = await this.ordersRepository.findOne({
+      where: { id: payload.idOrder },
+    });
+    if (!order) return null;
+    this.orderItemRepository.merge(orderItem, {
+      quantity: payload.quantity,
+      product,
+      order,
+    });
+    return await this.orderItemRepository.save(orderItem);
   }
 
+  async delete(idOrdenItem: number) {
+    const item = await this.orderItemRepository.findOne({
+      where: {
+        id: idOrdenItem,
+      },
+    });
+    if (!item) return null;
+    return await this.orderItemRepository.remove(item);
+  }
 }
